implement hunger process

+++ src/main.rs
@@ -164,7 +164,7 @@ fn main() {
         let color = [1.0, 1.0, 1.0];
         components.new_render_component(entity, Pix::Dood, color);
         components.new_position_component(entity, loc);
-        components.new_hunger_component(entity, 100.0);
+        components.new_hunger_component(entity, 100 as u16);
         chunk.insert_entity(entity); // maybe chunks.add_entity_to_chunk?
         // TODO do bounds checking
         scratch.insert_into_entities(entity);
